More modules! - Initial pass of modularization of quaternion and matrix libraries

+++ Slimetasia/RenderGBuffer.h
@@ -14,8 +14,8 @@ public:
 
     void SetWindowExtent(const vk::Extent2D& extent) override;
 
-    static constexpr size_t MAX_OBJECTS = 1000;
-    static constexpr size_t MAX_BONES = 32;
+    static constexpr uint32_t MAX_OBJECTS = 1000;
+    static constexpr uint32_t MAX_BONES = 32;
 
 protected:
